requirements.txt
firebase-admin
flask-restful
openai
python-dotenv
gunicorn